refactor(automation): extract schedule controls
